requirements.txt
praw
edge-tts
faster-whisper
moviepy
PyYAML
tenacity
